+++ src/OneSignalFacade.php
@@ -12,6 +12,10 @@ use Illuminate\Support\Facades\Facade;
  */
 class OneSignalFacade extends Facade
 {
+    /**
+     * getFacadeAccessor
+     * @return string
+     */
     protected static function getFacadeAccessor()
     {
         return 'onesignal';