+++ CHANGELOG.md
@@ -4,5 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.0.0]
-- Initial version
+## [1.0.0] - 2019-11-29
+### Added
+- Send push notifications for all platforms
+- Send push notifications for specific segments
+- Send push notifications for individual users
+- Send push notifications for iOS only
+- Send push notifications for Android only
+- Send push notifications for Web only
+- Push with specific content [see official documentation](https://documentation.onesignal.com/reference#create-notification)
+- Get the all push notifications
+- Get the push notifications specific sent by dashboard, api or automated
+- Get individual push notification
+- Delete a push notification
+- Get all players